+++ substratevm/src/com.oracle.svm.interpreter.metadata/src/com/oracle/svm/interpreter/metadata/InterpreterResolvedJavaMethod.java
@@ -41,12 +41,14 @@ import java.lang.reflect.Modifier;
 import java.lang.reflect.Type;
 import java.util.HashSet;
 import java.util.Set;
+import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 import java.util.function.Function;
 
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
 
 import com.oracle.svm.core.FunctionPointerHolder;
+import com.oracle.svm.core.SubstrateMetadata;
 import com.oracle.svm.core.hub.RuntimeClassLoading;
 import com.oracle.svm.core.hub.crema.CremaSupport;
 import com.oracle.svm.core.hub.registry.SymbolsSupport;
@@ -54,6 +56,7 @@ import com.oracle.svm.core.invoke.ResolvedMember;
 import com.oracle.svm.core.invoke.Target_java_lang_invoke_MemberName;
 import com.oracle.svm.core.meta.MethodPointer;
 import com.oracle.svm.core.util.VMError;
+import com.oracle.svm.espresso.classfile.Constants;
 import com.oracle.svm.espresso.classfile.JavaVersion;
 import com.oracle.svm.espresso.classfile.ParserMethod;
 import com.oracle.svm.espresso.classfile.attributes.CodeAttribute;
@@ -84,7 +87,7 @@ import jdk.vm.ci.meta.annotation.AnnotationsInfo;
  * Encapsulates resolved methods used under close-world assumptions, compiled and interpretable, but
  * also abstract methods for vtable calls.
  */
-public class InterpreterResolvedJavaMethod extends InterpreterAnnotated implements ResolvedJavaMethod, CremaMethodAccess, ResolvedMember {
+public class InterpreterResolvedJavaMethod extends InterpreterAnnotated implements ResolvedJavaMethod, CremaMethodAccess, ResolvedMember, SubstrateMetadata {
     @Platforms(Platform.HOSTED_ONLY.class)//
     @SuppressWarnings("unchecked") //
     private static final Class<? extends Annotation> CALLER_SENSITIVE_CLASS = (Class<? extends Annotation>) ReflectionUtil.lookupClass("jdk.internal.reflect.CallerSensitive");
@@ -132,6 +135,11 @@ public class InterpreterResolvedJavaMethod extends InterpreterAnnotated implemen
 
     // Token set by the toggle of method enter/exit events.
     private volatile Object interpreterExecToken;
+
+    // TODO move to crema once GR-71517 is resolved
+    private volatile ResolvedJavaMethod ristrettoMethod;
+    private static final AtomicReferenceFieldUpdater<InterpreterResolvedJavaMethod, ResolvedJavaMethod> RISTRETTO_METHOD_UPDATER = AtomicReferenceFieldUpdater
+                    .newUpdater(InterpreterResolvedJavaMethod.class, ResolvedJavaMethod.class, "ristrettoMethod");
 
     public static class InlinedBy {
         public InterpreterResolvedJavaMethod holder;
@@ -302,6 +310,27 @@ public class InterpreterResolvedJavaMethod extends InterpreterAnnotated implemen
         int flags = createFlags(modifiers, declaringClass, signatureSymbol, isSubstitutedNative, originalMethod);
         return new InterpreterResolvedJavaMethod(originalMethod, nameSymbol, maxLocals, maxStackSize, flags, declaringClass, signature, signatureSymbol, code,
                         exceptionHandlers, lineNumberTable, localVariableTable, nativeEntryPoint, vtableIndex, gotOffset, enterStubOffset, methodId);
+    }
+
+    public ResolvedJavaMethod getRistrettoMethod(Function<InterpreterResolvedJavaMethod, ResolvedJavaMethod> ristrettoMethodSupplier) {
+        if (this.ristrettoMethod != null) {
+            return this.ristrettoMethod;
+        }
+        /*
+         * We allow concurrent allocation of a ristretto method per interpreter method. Eventually
+         * however we CAS on the pointer in the interpreter representation, if another thread was
+         * faster return its method.
+         */
+        return getOrSetRistrettoMethod(ristrettoMethodSupplier.apply(this));
+    }
+
+    private ResolvedJavaMethod getOrSetRistrettoMethod(ResolvedJavaMethod newRistrettoMethod) {
+        if (RISTRETTO_METHOD_UPDATER.compareAndSet(this, null, newRistrettoMethod)) {
+            return newRistrettoMethod;
+        }
+        var method = this.ristrettoMethod;
+        assert method != null : "If CAS for null fails must have written a method already";
+        return method;
     }
 
     private static Symbol<Signature> toSymbol(InterpreterUnresolvedSignature jvmciSignature) {
@@ -803,7 +832,7 @@ public class InterpreterResolvedJavaMethod extends InterpreterAnnotated implemen
 
     @Override
     public final boolean isBridge() {
-        throw VMError.intentionallyUnimplemented();
+        return (Constants.ACC_BRIDGE & getModifiers()) != 0;
     }
 
     @Override
@@ -818,7 +847,8 @@ public class InterpreterResolvedJavaMethod extends InterpreterAnnotated implemen
 
     @Override
     public final StackTraceElement asStackTraceElement(int bci) {
-        throw VMError.intentionallyUnimplemented();
+        // TODO GR-71255
+        return new StackTraceElement(getDeclaringClass().getName(), getName(), null, bci);
     }
 
     @Override